requirements.txt
openai
edge-tts
openai-whisper
python-telegram-bot
